refactor: Extract the npm tarball reader into a shared helper
Move the minimal ustar reader out of agent-sdk-assets.ts into
npm-tarball.ts with a parameterized gzip-bomb cap, and share the spec
tarball fixtures via __test-helpers__. Also export isCompiledBinary so
the upcoming runtime bundle downloader can gate on the same detection.
No behavior change.
